requirements.txt
streamlit
pandas
Pillow
boto3
lxml_html_clean
uuid
botocore
chardet
pymupdf